Travis CI: add sanity check for libelf dependency

+++ travis-ci/managers/xenial.sh
@@ -14,4 +14,9 @@ CFLAGS="-g -O2 -Werror -Wall -fsanitize=address,undefined"
 mkdir build
 cc --version
 make CFLAGS="${CFLAGS}" -C ./src -B OBJDIR=../build
+ldd build/libbpf.so
+if ! ldd build/libbpf.so | grep -q libelf; then
+    echo "FAIL: No reference to libelf.so in libbpf.so!"
+    exit 1
+fi
 rm -rf build
